Move languages into their own module
