+++ src/lib.rs
@@ -22,6 +22,9 @@ pub use crate::node::*;
 mod rca_metrics;
 pub(crate) use rca_metrics::*;
 
+mod rca_languages;
+pub(crate) use rca_languages::*;
+
 pub mod web;
 
 #[macro_use]
@@ -68,45 +71,3 @@ pub use crate::comment_rm::*;
 
 mod dump;
 pub use crate::dump::*;
-
-mod language_ccomment;
-pub use crate::language_ccomment::*;
-
-mod language_cpp;
-pub use crate::language_cpp::*;
-
-mod language_csharp;
-pub use crate::language_csharp::*;
-
-mod language_css;
-pub use crate::language_css::*;
-
-mod language_go;
-pub use crate::language_go::*;
-
-mod language_html;
-pub use crate::language_html::*;
-
-mod language_java;
-pub use crate::language_java::*;
-
-mod language_mozjs;
-pub use crate::language_mozjs::*;
-
-mod language_javascript;
-pub use crate::language_javascript::*;
-
-mod language_python;
-pub use crate::language_python::*;
-
-mod language_rust;
-pub use crate::language_rust::*;
-
-mod language_tsx;
-pub use crate::language_tsx::*;
-
-mod language_typescript;
-pub use crate::language_typescript::*;
-
-mod language_preproc;
-pub use crate::language_preproc::*;
